requirements.txt
streamlit
pytesseract
PyPDF2
requests
pillow
google-generativeai
python-dotenv